Add files via upload
You must run system1.py file which is connected to the system2.py travel.db is the database of the system. You get the login access of the user and result of the pickup point  graph. To run the program you must have PyQt5 and matplotlib package to installed on your system.